Moved member view logic into shared domain module
ref https://github.com/TryGhost/Ghost/pull/26617

Unified parsing and active-match logic across hooks and sidebar by extracting member view domain helpers and reusing them from both consumers.
